pyproject.toml
[project]
name = "telegram-llm-kit"
version = "0.1.0"
description = "A reference implementation for LLM + RAG with Telegram as UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "tiktoken>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "respx>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
